Add SEGGER RTT for fast debug logging
Since we don't have SWO debugging available on a M0+ core,
we need a way to get debug logs.

Semihosting is incredible slow, so we use the RTT implementation
of SEGGER. The software part is open source. The only hardware
that implements RTT are the J-Link programmer by SEGGER themself.