Fix "01:00" bug
The TimerB was set to 01:00, which resulted in halting the MCU
when going into standby when the TimerB was active.

I reworked the timer logic and now the MCU doesn't freeze anymore.
We don't need an interrupt, since our timer are only minute sensitive
and we wake up the MCU every minute anyhow.